+++ kalliope/stt/cmusphinx/cmusphinx.py
@@ -4,11 +4,11 @@ from kalliope.core import Utils
 from kalliope.stt.Utils import SpeechRecognition
 
 
-class Wit(SpeechRecognition):
+class Cmusphinx(SpeechRecognition):
 
     def __init__(self, callback=None, **kwargs):
         """
-        Start recording the microphone and analyse audio with Wit.ai api
+        Start recording the microphone and analyse audio with CMU sphinx api
         :param callback: The callback function to call to send the text
         :param kwargs:
         """
@@ -16,27 +16,26 @@ class Wit(SpeechRecognition):
 
         # callback function to call after the translation speech/tex
         self.main_controller_callback = callback
-        self.key = kwargs.get('key', None)
-        self.show_all = kwargs.get('show_all', False)
 
         # start listening in the background
-        self.set_callback(self.wit_callback)
+        self.set_callback(self.sphinx_callback)
         self.start_listening()
 
-    def wit_callback(self, recognizer, audio):
+    def sphinx_callback(self, recognizer, audio):
+        """
+        called from the background thread
+        """
         try:
-            captured_audio = recognizer.recognize_wit(audio,
-                                                      key=self.key,
-                                                      show_all=self.show_all)
-            Utils.print_success("Wit.ai Speech Recognition thinks you said %s" % captured_audio)
+            captured_audio = recognizer.recognize_sphinx(audio)
+            Utils.print_success("Sphinx Speech Recognition thinks you said %s" % captured_audio)
             self._analyse_audio(captured_audio)
 
         except sr.UnknownValueError:
-            Utils.print_warning("Wit.ai Speech Recognition could not understand audio")
+            Utils.print_warning("Sphinx Speech Recognition could not understand audio")
             # callback anyway, we need to listen again for a new order
             self._analyse_audio(audio_to_text=None)
         except sr.RequestError as e:
-            Utils.print_danger("Could not request results from Wit.ai Speech Recognition service; {0}".format(e))
+            Utils.print_danger("Could not request results from Sphinx Speech Recognition service; {0}".format(e))
             # callback anyway, we need to listen again for a new order
             self._analyse_audio(audio_to_text=None)